requirements.txt
chromadb
openai
ragas
sentence_transformers
langchain
langchain_ollama
langchain-huggingface
langchain_community
langchain_text_splitters
rapidfuzz
pypdf
asyncio
python-dotenv

